dando uniformidad al código, retorno de errores desde los casos de uso, consumidos desde el controlador

+++ src/domain/use-cases/course/delete-course.ts
@@ -1,3 +1,5 @@
+import { CourseModel } from "../../../data";
+import { CustomError } from "../../errors/custom-error";
 import { CourseRepository } from "../../repository/course-repository";
 
 export interface DeleteCourseUseCase {
@@ -10,8 +12,12 @@ export class DeleteCourse implements DeleteCourseUseCase {
         private readonly courseRepository : CourseRepository, 
     ) {}
 
-    execute(id: string): Promise<boolean> {
-        return this.courseRepository.deleteCourse( id );
+    async execute(id: string): Promise<boolean> {
+        const courseToRemove = await this.courseRepository.getCourseById( id );
+        if(courseToRemove === null) throw CustomError.notFound(`El curso con el id: ${id}, no fue encontrado`);
+
+        const hasRemoved = await this.courseRepository.deleteCourse( id );
+        return hasRemoved;
     }
 
 }
